Add percent_complete to bookmark schema
This is the first step in supporting reading progress. The rough
architecture I have in mind is:

* User scrolls the page
* Amount scrolled it picked up (debounced)
  * this will likely reuse or copy logic from
    `scroll_indicator_controller.js`
* Send request to update bookmark `reading_progress` via API
  * maybe the existing bookmark update, maybe a dedicated route if
    needed for performance
* When loading a bookmark, scroll to same progress loaded from the
  bookmark

## Next steps (in order)

- [x] Add `percent_complete` to bookmarks
- [ ] API support for updating progress (including updated docs)
- [ ] Page restores progress position
- [ ] Page updates position and remove any transitional docs because
everything is done now

## Implementation notes

I'm storing the `reading_progress` as a float from 0–1. It could also
be stored other ways (DOM location like highlights, word offset, maybe
others), but this seemed like it would be the most understandable and
accurate to store/retrieve.

I decided to mark any archived bookmark as having `reading_progress =
1` (100%). It's a bit arbitrary since you can archive in the middle of
reading, but since we don't have historic reading info, it seemed like
the best way to fake it.

This commit obviously doesn't _do_ much of anything from a user's
perspective. That's intentional. I'm quite novice at go so I figured it
was better to make small changes that can be deployed at any time than
try to do everything all at once and risk abandonment. At least this
way if I drop this there's work and a plan for someone else to build
on. It's also easier to review.

Finally, I'm very open to other approaches here, either technically or
strategically.