Improved the way the Host and URL are built when a request starts.
This creates a HostInfo instance that stores the host and scheme of a
request based on reverse proxy headers or the settings' base_url.

When a request appears to come from a reverse proxy the scheme defaults
to HTTPS (and later takes any value given by x-forwarded-proto). This
should fix a most common issue with onboarding and sign-in.

Regardless of the base_url setting, the client IP is determined by
x-forwarded-for if the proxy is to be trusted.

resolves #620