Documentation translation using gettext.
The script docs/messages.py can extract blocks from the source markdown
files (en-US) and generate a .pot file.

The generate command uses the same extraction logic and applies the
target's translation on each block using the gettext file.

The generated po files are based on the last synchronized translations.
English will be put up to date with the main branch to update all the
source and translations.