Fix browser & pocket importer tests by parsing timestamps as UTC
"browser" and "pocket-file" importers were using `time.Unix()` to parse
Unix timestamps, but that created Go time objects with a local time zone.
The tests were expecting time objects in UTC time, and were thus failing on my
machine which is in CEST right now. This change makes it so that importers parse
all timestamps as UTC instead of translating them as local time.